fix: move types to runtime entrypoint
